+++ packages/lumx-react/src/components/progress/ProgressLinear.stories.tsx
@@ -1,0 +1,8 @@
+import { ProgressLinear } from '@lumx/react';
+import React from 'react';
+
+export default {
+    title: 'LumX components/progress/ProgressLinear',
+};
+
+export const Default = ({ theme }: any) => <ProgressLinear theme={theme} />;