+++ packages/lumx-react/src/components/progress/index.ts
@@ -1,1 +1,3 @@
 export * from './Progress';
+export * from './ProgressCircular';
+export * from './ProgressLinear';