+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
--   New component in React `GridColumn`.
+-   New `GridColumn` component.
+-   Progress: split into `ProgressCircular` and `ProgressLinear` components. `Progress` is now deprecated.
+-   ProgressCircular: add `size` variants.
 
 ## [3.0.7][] - 2022-12-06